Added sign to Integer Types and updated Integer attribute and surrounding code to handle signed and unsigned integer types. Fixed implicit cast not being axposed as actual value after variable declaration. Improved error handling in deserializers.

+++ dgir/vm/src/dgir/vm/dialect/arith/NumericUtils.java
@@ -42,7 +42,7 @@ final class NumericUtils {
    */
   static @NotNull Number convertToType(@NotNull Number value, @NotNull Type targetType) {
     if (targetType instanceof IntegerT integerT) {
-      return integerT.convertToValidNumber(value);
+      return integerT.convertToValidNumber(value.longValue());
     }
     if (targetType instanceof FloatT floatT) {
       return floatT.convertToValidNumber(value);
